Fix major bug with using a hash key.
Everytime a batch wanted state from a RedisState with a defined hkey (ie, where the state was stored in a hash and not the main namespace), we were doing an hgetall, thereby transferring the entire hash key (and the entire state), then filtering through all of the state to just get the values for the keys that were in the batch. This results in massive network traffic, as well as additional work on the client side.

Instead, use the Redis commad hmget (analogous to mget, but for a given hash key). This means that the client will only request the state for the keys that are in the batch. Way, way faster and more efficient.